Modifying if conditions for better performance

+++ 0x03-debugging/2-main.c
@@ -8,14 +8,14 @@
 
 int main(void)
 {
-        float a, b, c;
-        float largest;
+        int a, b, c;
+        int largest;
 	
 	while (1)
 	{
-		scanf("%f %f %f", &a,&b,&c);
+		scanf("%d %d %d", &a,&b,&c);
         	largest = largest_number(a, b, c);
-        	printf("%.0f is the largest number\n", largest);
+        	printf("%d is the largest number\n", largest);
 	}
